Fix the race-condition fix, so it works in NRI CI
NoRedInk CI merges in master before running anything in CI. This means
that taking HEAD from git includes all master changes merged in.

This is what was causing the race condition fix in 0.13.3 to trigger
absurd checks for simple branches.

This commit introduces an option to override the head commit used in the
checks, so we can pass e.g. the GIT_COMMIT env var from Jenkins and make
the checks sane again.